requirements.txt
numpy
opencv-python
pillow
fastapi